backend/requirements.txt

fastapi
uvicorn[standard]


sqlalchemy
psycopg2-binary
alembic


pydantic
pydantic-settings


python-dotenv


passlib[bcrypt]
python-jose[cryptography]


google-api-python-client
google-auth-httplib2
google-auth-oauthlib


openai